refactor(xapiri): extract dashboard_fields.go — enums, fieldMeta, renderField
Part 2 of ADR 0014 dashboard split. Moves palette, all four enum namespaces
(tiCount/siCount/toiCount/focCount), fkind, fieldMeta, dashFields, selectState,
labelW/inputW, and renderField from the 4422-line dashboard.go into a dedicated
dashboard_fields.go.

Trap 1 compliance: renderField stays co-located with dashFields in the same file.
Trap 2 compliance: all iota enums remain in a single const block per namespace
in dashboard_fields.go; no iota appears in any tab_*.go file.
TestRenderField_SecretFieldsNeverLeakValue passes throughout.
